Refactor Mirage handlers into separate files
