+++ mirage/handlers/affiliate-earnings-payouts.js
@@ -1,0 +1,4 @@
+export default function (server) {
+  server.get('/affiliate-earnings-payouts');
+  server.post('/affiliate-earnings-payouts');
+}
