+++ mirage/handlers/affiliate-referrals.js
@@ -1,0 +1,8 @@
+export default function (server) {
+  server.post('/affiliate-referrals', function (schema) {
+    const attrs = this.normalizedRequestAttrs();
+    attrs.activatedAt = new Date();
+
+    return schema.affiliateReferrals.create(attrs);
+  });
+}
