+++ mirage/handlers/analytics-events.js
@@ -1,0 +1,3 @@
+export default function (server) {
+  server.post('/analytics-events');
+}
