+++ mirage/handlers/badges.js
@@ -1,0 +1,3 @@
+export default function (server) {
+  server.get('/badges');
+}
