+++ mirage/handlers/course-stage-completions.js
@@ -1,0 +1,8 @@
+export default function (server) {
+  server.post('/course-stage-completions', function (schema) {
+    const attrs = this.normalizedRequestAttrs();
+    attrs.completedAt = new Date();
+
+    return schema.courseStageCompletions.create(attrs);
+  });
+}
