+++ mirage/handlers/individual-payment-method-update-requests.js
@@ -1,0 +1,5 @@
+export default function (server) {
+  server.post('/individual-payment-method-update-requests', function (schema) {
+    return schema.individualPaymentMethodUpdateRequests.create({ url: 'https://test.com/checkout_session' });
+  });
+}
