+++ mirage/handlers/languages.js
@@ -1,0 +1,5 @@
+export default function (server) {
+  server.get('/languages', function (schema) {
+    return schema.languages.all();
+  });
+}
