+++ mirage/handlers/leaderboard-entries.js
@@ -1,0 +1,19 @@
+export default function (server) {
+  server.get('/course-leaderboard-entries', function (schema, request) {
+    let result = schema.courseLeaderboardEntries.all();
+
+    if (request.queryParams.team_id) {
+      const team = schema.teams.find(request.queryParams.team_id);
+      const teamMemberships = schema.teamMemberships.where({ teamId: team.id }).models;
+      const userIds = teamMemberships.map((teamMembership) => teamMembership.user.id);
+
+      result = result.filter((leaderboardEntry) => userIds.includes(leaderboardEntry.user.id));
+    }
+
+    if (request.queryParams.course_id) {
+      result = result.filter((leaderboardEntry) => leaderboardEntry.currentCourseStage.course.id === request.queryParams.course_id);
+    }
+
+    return result;
+  });
+}
