+++ mirage/handlers/sessions.js
@@ -1,0 +1,5 @@
+export default function (server) {
+  server.post('/sessions/logout', function () {
+    return new Response(200, {}, {});
+  });
+}
