+++ mirage/handlers/slack-integrations.js
@@ -1,0 +1,3 @@
+export default function (server) {
+  server.delete('/slack-integrations/:id');
+}
