+++ mirage/handlers/team-memberships.js
@@ -1,0 +1,3 @@
+export default function (server) {
+  server.delete('team-memberships/:id');
+}
