+++ mirage/handlers/team-payment-method-update-requests.js
@@ -1,0 +1,5 @@
+export default function (server) {
+  server.post('/team-payment-method-update-requests', function (schema) {
+    return schema.teamPaymentMethodUpdateRequests.create({ url: 'https://test.com/team_payment_method_update_request' });
+  });
+}
