+++ mirage/handlers/track-leaderboard-entries.js
@@ -1,0 +1,11 @@
+export default function (server) {
+  server.get('/track-leaderboard-entries', function (schema, request) {
+    let result = schema.trackLeaderboardEntries.all();
+
+    if (request.queryParams.language_id) {
+      result = result.filter((trackLeaderboardEntry) => trackLeaderboardEntry.language.id === request.queryParams.language_id);
+    }
+
+    return result;
+  });
+}
