+++ mirage/handlers/views.js
@@ -1,0 +1,3 @@
+export default function (server) {
+  server.post('/views');
+}
